requirements.txt
dify_plugin~=0.0.1b72
requests
opencv-python
numpy
Pillow
